pyproject.toml
[tool.poetry]
name = "smarthub"
version = "0.1.0"
description = ""
authors = ["sergey <sbitza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
fastapi-users = {extras = ["sqlalchemy"], version = "^14.0.1"}
pydantic-settings = "^2.8.0"
aio-pika = "^9.5.5"
paho-mqtt = "^2.1.0"
pytz = "^2025.1"
requests = "^2.32.3"
sqladmin = "^0.20.1"
sqlalchemy = "^2.0.39"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
itsdangerous = "^2.2.0"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
